moved validation to server, changed function name
